requirements-test.txt
# dependencies for test
pytest              # test framework
pytest-cov          # coverage test

boto3
attrs
jsonpickle